third_party/blink/web_tests/fast/forms/radio/radio-restore-preventDefault-expected.txt

PASS document.getElementById("radio1").checked is true
PASS document.getElementById("radio2").checked is false
PASS document.getElementById("radio1").checked is false
PASS document.getElementById("radio2").checked is false
PASS successfullyParsed is true

TEST COMPLETE

A B
